refactor: update imports and fix formatting
- Update file extensions from .js to .ts where applicable
- Fix spacing in type imports
- Standardize compose.yaml quotes
- Clean up README.md formatting
