+++ README.md
@@ -3,28 +3,28 @@
 The Ansopedia User Service is a backend service responsible for managing user accounts and authentication within the Ansopedia learning platform.  
 It provides functionalities like:
 
-- **User Registration and Login:** Create accounts and log in securely.  
-- **User Management:** Manage profiles, preferences, and roles.  
-- **Authentication:** Secure access with JWT-based authentication.  
+- **User Registration and Login:** Create accounts and log in securely.
+- **User Management:** Manage profiles, preferences, and roles.
+- **Authentication:** Secure access with JWT-based authentication.
 - **Integration:** Works with other services like Ansopedia Studio API to manage user permissions.
 
 ## Understanding the Scripts
 
-- **build:** Transpiles TypeScript and fixes path aliases.  
-- **local:** Starts the server with `NODE_ENV=local`.  
-- **local:dev:** Starts the server with `NODE_ENV=development`.  
-- **local:prod:** Starts the server with `NODE_ENV=production` and runs the production build.  
-- **generate-keys:** Generates RSA keys for JWT authentication.  
-- **generate-env:** Generates `.env` files from templates.  
-- **migrate:** Runs migrations for the current `NODE_ENV` (used internally by other migrate scripts).  
-- **migrate:local:** Runs migrations for `NODE_ENV=local`.  
-- **migrate:prod:** Runs migrations for `NODE_ENV=production`.  
-- **setup:** Generates environment files and RSA keys.  
-- **lint / lint:fix:** Lint code and optionally fix issues.  
-- **prettier:check / prettier:fix:** Check or fix formatting.  
-- **prepare:** Runs Husky hooks.  
-- **prod:** Builds the app and starts in production.  
-- **start:** Runs the built app.  
+- **build:** Transpiles TypeScript and fixes path aliases.
+- **local:** Starts the server with `NODE_ENV=local`.
+- **local:dev:** Starts the server with `NODE_ENV=development`.
+- **local:prod:** Starts the server with `NODE_ENV=production` and runs the production build.
+- **generate-keys:** Generates RSA keys for JWT authentication.
+- **generate-env:** Generates `.env` files from templates.
+- **migrate:** Runs migrations for the current `NODE_ENV` (used internally by other migrate scripts).
+- **migrate:local:** Runs migrations for `NODE_ENV=local`.
+- **migrate:prod:** Runs migrations for `NODE_ENV=production`.
+- **setup:** Generates environment files and RSA keys.
+- **lint / lint:fix:** Lint code and optionally fix issues.
+- **prettier:check / prettier:fix:** Check or fix formatting.
+- **prepare:** Runs Husky hooks.
+- **prod:** Builds the app and starts in production.
+- **start:** Runs the built app.
 - **test / test:coverage:** Run tests, optionally with coverage.
 
 ## Project Setup
@@ -51,7 +51,6 @@ Follow these steps to set up the project:
    ```
 
    This:
-
    - Creates `.env` files for local, development, and test.
    - Generates RSA keys.
    - Injects keys into `.env` files.
@@ -91,36 +90,30 @@ Follow these steps to set up the project:
 ## Available Scripts
 
 - **Setup:**
-
   - `pnpm setup` → Generate env files & RSA keys.
   - `pnpm generate-env` → Env files only.
   - `pnpm generate-keys` → RSA keys only.
 
 - **Migrations:**
-
   - `pnpm migrate` → Run for current `NODE_ENV`.
   - `pnpm migrate:local` → Run for local.
   - `pnpm migrate:prod` → Run for production.
 
 - **Development:**
-
   - `pnpm local` → Local environment.
   - `pnpm local:dev` → Development environment.
   - `pnpm local:prod` → Production environment (locally).
 
 - **Production:**
-
   - `pnpm build` → Build app.
   - `pnpm prod` → Build & run.
   - `pnpm start` → Start built app.
 
 - **Testing:**
-
   - `pnpm test`
   - `pnpm test:coverage`
 
 - **Code Quality:**
-
   - `pnpm lint`
   - `pnpm lint:fix`
   - `pnpm prettier:check`
@@ -154,29 +147,29 @@ You should pass environment variables like `NODE_ENV` and any secrets (database 
 
 Example using environment variables inline:
 
-  ```bash
-  docker run -d -p 3000:3000 \
-    -e NODE_ENV=production \
-    -e DATABASE_URI=your_DATABASE_URI \
-    -e JWT_PRIVATE_KEY="your_jwt_private_key" \
-    ansopedia-user-service
-  ```
+```bash
+docker run -d -p 3000:3000 \
+  -e NODE_ENV=production \
+  -e DATABASE_URI=your_DATABASE_URI \
+  -e JWT_PRIVATE_KEY="your_jwt_private_key" \
+  ansopedia-user-service
+```
 
 Alternatively, use an `.env` file to manage environment variables:
 
 1. Create a `.env` file with the needed variables:
 
-    ```env
-      NODE_ENV=production
-      DATABASE_URI=your_DATABASE_URI
-      JWT_PRIVATE_KEY="your_jwt_private_key"
-    ```
+   ```env
+     NODE_ENV=production
+     DATABASE_URI=your_DATABASE_URI
+     JWT_PRIVATE_KEY="your_jwt_private_key"
+   ```
 
 2. Run the container with the env file:
 
-    ```bash
-    docker run -d -p 3000:3000 --env-file .env ansopedia-user-service
-    ```
+   ```bash
+   docker run -d -p 3000:3000 --env-file .env ansopedia-user-service
+   ```
 
 ### Notes
