+++ src/api/v1/profile/__test__/profle.test.ts
@@ -1,5 +1,5 @@
 import { defaultUsers } from "@/constants";
-import {type  MongooseObjectId } from "@/types";
+import { type MongooseObjectId } from "@/types";
 import {
   expectBadRequestResponseForValidationError,
   expectLoginSuccess,
