+++ src/middlewares/auth.ts
@@ -2,7 +2,7 @@ import type { NextFunction, Request, Response } from "express";
 
 import { AuthService } from "@/api/v1/auth/auth.service.js";
 import { ErrorTypeEnum } from "@/constants";
-import { extractTokenFromBearerString } from "@/utils/jwt.util.js";
+import { extractTokenFromBearerString } from "@/utils";
 
 export const authenticate = async (req: Request, res: Response, next: NextFunction) => {
   try {
